+++ server/package.json
@@ -43,6 +43,7 @@
     "@types/morgan": "^1.9.4",
     "@types/multer": "^1.4.7",
     "@types/node": "^20.3.2",
+    "@types/node-cache": "^4.2.5",
     "ts-node": "^10.9.1",
     "typescript": "^5.1.6"
   }
